+++ source/ftrack_application_launcher/configure_logging.py
@@ -113,4 +113,4 @@ def configure_logging(logger_name, level=None, format=None, extra_modules=None):
     logging.captureWarnings(True)
 
     # Log out the file output.
-    logging.warning('Saving log file to: {0}'.format(logfile))
+    logging.info('Saving log file to: {0}'.format(logfile))