+++ source/ftrack_application_launcher/usage.py
@@ -26,39 +26,45 @@ def get_session():
     )
 
 
-def _send_event(event_name, metadata=None):
+def _send_event(session, event_name, metadata=None):
     '''Send usage event with *event_name* and *metadata*.'''
-    global _log_usage_session
 
-    if _log_usage_session is None:
-        _log_usage_session = get_session()
+    if not isinstance(metadata, list):
+        metadata = [metadata]
+
+    payload = []
+    for data in metadata:
+        payload.append(
+            {
+                'action': '_track_usage',
+                'data': {
+                    'type': 'event',
+                    'name': event_name,
+                    'metadata': data
+                }
+            }
+        )
 
     try:
-        _log_usage_session.call([{
-            'action': '_track_usage',
-            'data': {
-                'type': 'event',
-                'name': event_name,
-                'metadata': metadata
-            }
-        }])
+        session.call(payload)
+
     except Exception:
         logger.exception('Failed to send event : {}'.format(event_name))
 
 
 @asynchronous.asynchronous
-def _send_async_event(event_name, metadata=None):
+def _send_async_event(session, event_name, metadata=None):
     '''Call __send_event in a new thread.'''
-    _send_event(event_name, metadata)
+    _send_event(session, event_name, metadata)
 
 
-def send_event(event_name, metadata=None, asynchronous=True):
+def send_event(session, event_name, metadata=None, asynchronous=True):
     '''Send usage event with *event_name* and *metadata*.
 
     If asynchronous is True, the event will be sent in a new thread.
     '''
 
     if asynchronous:
-        _send_async_event(event_name, metadata)
+        _send_async_event(session, event_name, metadata)
     else:
-        _send_event(event_name, metadata)
+        _send_event(session, event_name, metadata)